+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientManager.java
@@ -1,8 +1,12 @@
 package de.maxhenkel.voicechat.voice.client;
 
 import de.maxhenkel.voicechat.Voicechat;
+import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
+import de.maxhenkel.voicechat.macos.PermissionCheck;
+import de.maxhenkel.voicechat.macos.VersionCheck;
+import de.maxhenkel.voicechat.macos.avfoundation.AVAuthorizationStatus;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.net.RequestSecretPacket;
 import de.maxhenkel.voicechat.net.SecretPacket;
@@ -30,6 +34,7 @@ public class ClientManager {
     private final RenderEvents renderEvents;
     private final KeyEvents keyEvents;
     private final Minecraft minecraft;
+    private boolean hasShownPermissionsMessage;
 
     private ClientManager() {
         playerStateManager = new ClientPlayerStateManager();
@@ -87,9 +92,32 @@ public class ClientManager {
             Voicechat.LOGGER.info("Disconnecting from previous connection due to server change");
             onDisconnect();
         }
+        hasShownPermissionsMessage = false;
         Voicechat.LOGGER.info("Sending secret request to the server");
         NetManager.sendToServer(new RequestSecretPacket(Voicechat.COMPATIBILITY_VERSION));
         client = new ClientVoicechat();
+    }
+
+    public void checkMicrophonePermissions() {
+        if (!VoicechatClient.CLIENT_CONFIG.macosCheckMicrophonePermission.get()) {
+            return;
+        }
+        if (VersionCheck.isMacOSNativeCompatible()) {
+            AVAuthorizationStatus status = PermissionCheck.getMicrophonePermissions();
+            if (status.equals(AVAuthorizationStatus.DENIED)) {
+                if (!hasShownPermissionsMessage) {
+                    ClientManager.sendPlayerError("message.voicechat.macos_no_mic_permission", null);
+                    hasShownPermissionsMessage = true;
+                }
+                Voicechat.LOGGER.warn("User hasn't granted microphone permissions: {}", status.name());
+            } else if (!status.equals(AVAuthorizationStatus.AUTHORIZED)) {
+                if (!hasShownPermissionsMessage) {
+                    ClientManager.sendPlayerError("message.voicechat.macos_unsupported_launcher", null);
+                    hasShownPermissionsMessage = true;
+                }
+                Voicechat.LOGGER.warn("User has an unsupported launcher: {}", status.name());
+            }
+        }
     }
 
     public static void sendPlayerError(String translationKey, @Nullable Exception e) {